Added support for functions and constants in aggregate expression
