+++ src/Microsoft.AspNet.OData.Shared/Query/Expressions/ExpressionBinderBase.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.  See License.txt in the project root for license information.
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 #if NETFX // System.Data.Linq.Binary is only supported in the AspNet version.
 using System.Data.Linq;
@@ -27,6 +28,7 @@ namespace Microsoft.AspNet.OData.Query.Expressions
     /// <summary>
     /// The base class for all expression binders.
     /// </summary>
+    [SuppressMessage("Microsoft.Maintainability", "CA1506:AvoidExcessiveClassCoupling", Justification = "Relies on many ODataLib classes.")]
     public abstract class ExpressionBinderBase
     {
         internal static readonly MethodInfo StringCompareMethodInfo = typeof(string).GetMethod("Compare", new[] { typeof(string), typeof(string), typeof(StringComparison) });
@@ -1078,6 +1080,754 @@ namespace Microsoft.AspNet.OData.Query.Expressions
             }
 
             return 0;
+        }
+
+        /// <summary>
+        /// Binds a <see cref="QueryNode"/> to create a LINQ <see cref="Expression"/> that represents the semantics
+        /// of the <see cref="QueryNode"/>.
+        /// </summary>
+        /// <param name="node">The node to bind.</param>
+        /// <returns>The LINQ <see cref="Expression"/> created.</returns>
+        [SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity",
+            Justification = "These are simple conversion function and cannot be split up.")]
+        public abstract Expression Bind(QueryNode node);
+
+        /// <summary>
+        /// Gets $it parameter
+        /// </summary>
+        /// <returns></returns>
+        protected abstract ParameterExpression Parameter { get; }
+
+        /// <summary>
+        /// Binds a <see cref="ConstantNode"/> to create a LINQ <see cref="Expression"/> that
+        /// represents the semantics of the <see cref="ConstantNode"/>.
+        /// </summary>
+        /// <param name="constantNode">The node to bind.</param>
+        /// <returns>The LINQ <see cref="Expression"/> created.</returns>
+        public virtual Expression BindConstantNode(ConstantNode constantNode)
+        {
+            Contract.Assert(constantNode != null);
+
+            // no need to parameterize null's as there cannot be multiple values for null.
+            if (constantNode.Value == null)
+            {
+                return NullConstant;
+            }
+
+            object value = constantNode.Value;
+            Type constantType = RetrieveClrTypeForConstant(constantNode.TypeReference, ref value);
+
+            if (QuerySettings.EnableConstantParameterization)
+            {
+                return LinqParameterContainer.Parameterize(constantType, value);
+            }
+            else
+            {
+                return Expression.Constant(value, constantType);
+            }
+        }
+
+        /// <summary>
+        /// Binds a <see cref="SingleValueFunctionCallNode"/> to create a LINQ <see cref="Expression"/> that
+        /// represents the semantics of the <see cref="SingleValueFunctionCallNode"/>.
+        /// </summary>
+        /// <param name="node">The node to bind.</param>
+        /// <returns>The LINQ <see cref="Expression"/> created.</returns>
+        [SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity",
+                        Justification = "These are simple binding functions and cannot be split up.")]
+        public virtual Expression BindSingleValueFunctionCallNode(SingleValueFunctionCallNode node)
+        {
+            switch (node.Name)
+            {
+                case ClrCanonicalFunctions.StartswithFunctionName:
+                    return BindStartsWith(node);
+
+                case ClrCanonicalFunctions.EndswithFunctionName:
+                    return BindEndsWith(node);
+
+                case ClrCanonicalFunctions.ContainsFunctionName:
+                    return BindContains(node);
+
+                case ClrCanonicalFunctions.SubstringFunctionName:
+                    return BindSubstring(node);
+
+                case ClrCanonicalFunctions.LengthFunctionName:
+                    return BindLength(node);
+
+                case ClrCanonicalFunctions.IndexofFunctionName:
+                    return BindIndexOf(node);
+
+                case ClrCanonicalFunctions.TolowerFunctionName:
+                    return BindToLower(node);
+
+                case ClrCanonicalFunctions.ToupperFunctionName:
+                    return BindToUpper(node);
+
+                case ClrCanonicalFunctions.TrimFunctionName:
+                    return BindTrim(node);
+
+                case ClrCanonicalFunctions.ConcatFunctionName:
+                    return BindConcat(node);
+
+                case ClrCanonicalFunctions.YearFunctionName:
+                case ClrCanonicalFunctions.MonthFunctionName:
+                case ClrCanonicalFunctions.DayFunctionName:
+                    return BindDateRelatedProperty(node); // Date & DateTime & DateTimeOffset
+
+                case ClrCanonicalFunctions.HourFunctionName:
+                case ClrCanonicalFunctions.MinuteFunctionName:
+                case ClrCanonicalFunctions.SecondFunctionName:
+                    return BindTimeRelatedProperty(node); // TimeOfDay & DateTime & DateTimeOffset
+
+                case ClrCanonicalFunctions.FractionalSecondsFunctionName:
+                    return BindFractionalSeconds(node);
+
+                case ClrCanonicalFunctions.RoundFunctionName:
+                    return BindRound(node);
+
+                case ClrCanonicalFunctions.FloorFunctionName:
+                    return BindFloor(node);
+
+                case ClrCanonicalFunctions.CeilingFunctionName:
+                    return BindCeiling(node);
+
+                case ClrCanonicalFunctions.CastFunctionName:
+                    return BindCastSingleValue(node);
+
+                case ClrCanonicalFunctions.IsofFunctionName:
+                    return BindIsOf(node);
+
+                case ClrCanonicalFunctions.DateFunctionName:
+                    return BindDate(node);
+
+                case ClrCanonicalFunctions.TimeFunctionName:
+                    return BindTime(node);
+
+                case ClrCanonicalFunctions.NowFunctionName:
+                    return BindNow(node);
+
+                default:
+                    // Get Expression of custom binded method.
+                    Expression expression = BindCustomMethodExpressionOrNull(node);
+                    if (expression != null)
+                    {
+                        return expression;
+                    }
+
+                    throw new NotImplementedException(Error.Format(SRResources.ODataFunctionNotSupported, node.Name));
+            }
+        }
+
+        /// <summary>
+        /// Binds a <see cref="CollectionConstantNode"/> to create a LINQ <see cref="Expression"/> that
+        /// represents the semantics of the <see cref="CollectionConstantNode"/>.
+        /// </summary>
+        /// <param name="node">The node to bind.</param>
+        /// <returns>The LINQ <see cref="Expression"/> created.</returns>
+        public virtual Expression BindCollectionConstantNode(CollectionConstantNode node)
+        {
+            // It's fine if the collection is empty; the returned value will be an empty list.
+            ConstantNode firstNode = node.Collection.FirstOrDefault();
+            object value = null;
+            if (firstNode != null)
+            {
+                value = firstNode.Value;
+            }
+
+            Type constantType = RetrieveClrTypeForConstant(node.ItemType, ref value);
+            Type listType = typeof(List<>).MakeGenericType(constantType);
+            IList castedList = Activator.CreateInstance(listType) as IList;
+
+            // Getting a LINQ expression to dynamically cast each item in the Collection during runtime is tricky,
+            // so using a foreach loop and doing an implicit cast from object to the CLR type of ItemType.
+            foreach (ConstantNode item in node.Collection)
+            {
+                castedList.Add(item.Value);
+            }
+
+            return Expression.Constant(castedList);
+        }
+
+        private Expression BindCastSingleValue(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert(ClrCanonicalFunctions.CastFunctionName == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            Contract.Assert(arguments.Length == 1 || arguments.Length == 2);
+
+            Expression source = arguments.Length == 1 ? this.Parameter : arguments[0];
+            string targetTypeName = (string)((ConstantNode)node.Parameters.Last()).Value;
+            IEdmType targetEdmType = Model.FindType(targetTypeName);
+            Type targetClrType = null;
+
+            if (targetEdmType != null)
+            {
+                IEdmTypeReference targetEdmTypeReference = targetEdmType.ToEdmTypeReference(false);
+                targetClrType = EdmLibHelpers.GetClrType(targetEdmTypeReference, Model);
+
+                if (source != NullConstant)
+                {
+                    if (source.Type == targetClrType)
+                    {
+                        return source;
+                    }
+
+                    if ((!targetEdmTypeReference.IsPrimitive() && !targetEdmTypeReference.IsEnum()) ||
+                        (EdmLibHelpers.GetEdmPrimitiveTypeOrNull(source.Type) == null && !TypeHelper.IsEnum(source.Type)))
+                    {
+                        // Cast fails and return null.
+                        return NullConstant;
+                    }
+                }
+            }
+
+            if (targetClrType == null || source == NullConstant)
+            {
+                return NullConstant;
+            }
+
+            if (targetClrType == typeof(string))
+            {
+                return BindCastToStringType(source);
+            }
+            else if (TypeHelper.IsEnum(targetClrType))
+            {
+                return BindCastToEnumType(source.Type, targetClrType, node.Parameters.First(), arguments.Length);
+            }
+            else
+            {
+                if (TypeHelper.IsNullable(source.Type) && !TypeHelper.IsNullable(targetClrType))
+                {
+                    // Make the target Clr type nullable to avoid failure while casting
+                    // nullable source, whose value may be null, to a non-nullable type.
+                    // For example: cast(NullableInt32Property,Edm.Int64)
+                    // The target Clr type should be Nullable<Int64> rather than Int64.
+                    targetClrType = typeof(Nullable<>).MakeGenericType(targetClrType);
+                }
+
+                try
+                {
+                    return Expression.Convert(source, targetClrType);
+                }
+                catch (InvalidOperationException)
+                {
+                    // Cast fails and return null.
+                    return NullConstant;
+                }
+            }
+        }
+
+        private static Expression BindCastToStringType(Expression source)
+        {
+            Expression sourceValue;
+
+            if (TypeHelper.IsGenericType(source.Type) && source.Type.GetGenericTypeDefinition() == typeof(Nullable<>))
+            {
+                if (TypeHelper.IsEnum(source.Type))
+                {
+                    // Entity Framework doesn't have ToString method for enum types.
+                    // Convert enum types to their underlying numeric types.
+                    sourceValue = Expression.Convert(
+                        Expression.Property(source, "Value"),
+                        Enum.GetUnderlyingType(TypeHelper.GetUnderlyingTypeOrSelf(source.Type)));
+                }
+                else
+                {
+                    // Entity Framework has ToString method for numeric types.
+                    sourceValue = Expression.Property(source, "Value");
+                }
+
+                // Entity Framework doesn't have ToString method for nullable numeric types.
+                // Call ToString method on non-nullable numeric types.
+                return Expression.Condition(
+                    Expression.Property(source, "HasValue"),
+                    Expression.Call(sourceValue, "ToString", typeArguments: null, arguments: null),
+                    Expression.Constant(null, typeof(string)));
+            }
+            else
+            {
+                sourceValue = TypeHelper.IsEnum(source.Type) ?
+                    Expression.Convert(source, Enum.GetUnderlyingType(source.Type)) :
+                    source;
+                return Expression.Call(sourceValue, "ToString", typeArguments: null, arguments: null);
+            }
+        }
+
+        private Expression BindCastToEnumType(Type sourceType, Type targetClrType, QueryNode firstParameter, int parameterLength)
+        {
+            Type enumType = TypeHelper.GetUnderlyingTypeOrSelf(targetClrType);
+            ConstantNode sourceNode = firstParameter as ConstantNode;
+
+            if (parameterLength == 1 || sourceNode == null || sourceType != typeof(string))
+            {
+                // We only support to cast Enumeration type from constant string now,
+                // because LINQ to Entities does not recognize the method Enum.TryParse.
+                return NullConstant;
+            }
+            else
+            {
+                object[] parameters = new[] { sourceNode.Value, Enum.ToObject(enumType, 0) };
+                bool isSuccessful = (bool)EnumTryParseMethod.MakeGenericMethod(enumType).Invoke(null, parameters);
+
+                if (isSuccessful)
+                {
+                    if (QuerySettings.EnableConstantParameterization)
+                    {
+                        return LinqParameterContainer.Parameterize(targetClrType, parameters[1]);
+                    }
+                    else
+                    {
+                        return Expression.Constant(parameters[1], targetClrType);
+                    }
+                }
+                else
+                {
+                    return NullConstant;
+                }
+            }
+        }
+
+        private Expression BindIsOf(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert(ClrCanonicalFunctions.IsofFunctionName == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+
+            // Edm.Boolean isof(type)  or
+            // Edm.Boolean isof(expression,type)
+            Contract.Assert(arguments.Length == 1 || arguments.Length == 2);
+
+            Expression source = arguments.Length == 1 ? this.Parameter : arguments[0];
+            if (source == NullConstant)
+            {
+                return FalseConstant;
+            }
+
+            string typeName = (string)((ConstantNode)node.Parameters.Last()).Value;
+
+            IEdmType edmType = Model.FindType(typeName);
+            Type clrType = null;
+            if (edmType != null)
+            {
+                // bool nullable = source.Type.IsNullable();
+                IEdmTypeReference edmTypeReference = edmType.ToEdmTypeReference(false);
+                clrType = EdmLibHelpers.GetClrType(edmTypeReference, Model);
+            }
+
+            if (clrType == null)
+            {
+                return FalseConstant;
+            }
+
+            bool isSourcePrimitiveOrEnum = EdmLibHelpers.GetEdmPrimitiveTypeOrNull(source.Type) != null ||
+                                           TypeHelper.IsEnum(source.Type);
+
+            bool isTargetPrimitiveOrEnum = EdmLibHelpers.GetEdmPrimitiveTypeOrNull(clrType) != null ||
+                                           TypeHelper.IsEnum(clrType);
+
+            if (isSourcePrimitiveOrEnum && isTargetPrimitiveOrEnum)
+            {
+                if (TypeHelper.IsNullable(source.Type))
+                {
+                    clrType = TypeHelper.ToNullable(clrType);
+                }
+            }
+
+            // Be caution: Type method of LINQ to Entities only supports entity type.
+            return Expression.Condition(Expression.TypeIs(source, clrType), TrueConstant, FalseConstant);
+        }
+
+        private Expression BindCeiling(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("ceiling" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+
+            Contract.Assert(arguments.Length == 1 && IsDoubleOrDecimal(arguments[0].Type));
+
+            MethodInfo ceiling = IsType<double>(arguments[0].Type)
+                ? ClrCanonicalFunctions.CeilingOfDouble
+                : ClrCanonicalFunctions.CeilingOfDecimal;
+            return MakeFunctionCall(ceiling, arguments);
+        }
+
+        private Expression BindFloor(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("floor" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+
+            Contract.Assert(arguments.Length == 1 && IsDoubleOrDecimal(arguments[0].Type));
+
+            MethodInfo floor = IsType<double>(arguments[0].Type)
+                ? ClrCanonicalFunctions.FloorOfDouble
+                : ClrCanonicalFunctions.FloorOfDecimal;
+            return MakeFunctionCall(floor, arguments);
+        }
+
+        private Expression BindRound(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("round" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+
+            Contract.Assert(arguments.Length == 1 && IsDoubleOrDecimal(arguments[0].Type));
+
+            MethodInfo round = IsType<double>(arguments[0].Type)
+                ? ClrCanonicalFunctions.RoundOfDouble
+                : ClrCanonicalFunctions.RoundOfDecimal;
+            return MakeFunctionCall(round, arguments);
+        }
+
+        private Expression BindDate(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("date" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+
+            // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
+            Contract.Assert(arguments.Length == 1 && IsDateOrOffset(arguments[0].Type));
+
+            // EF doesn't support new Date(int, int, int), also doesn't support other property access, for example DateTime.Date.
+            // Therefore, we just return the source (DateTime or DateTimeOffset).
+            return arguments[0];
+        }
+
+        private Expression BindNow(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("now" == node.Name);
+
+            // Function Now() does not take any arguemnts.
+            Expression[] arguments = BindArguments(node.Parameters);
+            Contract.Assert(arguments.Length == 0);
+
+            return Expression.Property(null, typeof(DateTimeOffset), "UtcNow");
+        }
+
+        private Expression BindTime(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("time" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+
+            // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
+            Contract.Assert(arguments.Length == 1 && IsDateOrOffset(arguments[0].Type));
+
+            // EF doesn't support new TimeOfDay(int, int, int, int), also doesn't support other property access, for example DateTimeOffset.DateTime.
+            // Therefore, we just return the source (DateTime or DateTimeOffset).
+            return arguments[0];
+        }
+
+        private Expression BindFractionalSeconds(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("fractionalseconds" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            Contract.Assert(arguments.Length == 1 && (IsTimeRelated(arguments[0].Type)));
+
+            // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
+            Expression parameter = arguments[0];
+
+            PropertyInfo property;
+            if (IsTimeOfDay(parameter.Type))
+            {
+                property = ClrCanonicalFunctions.TimeOfDayProperties[ClrCanonicalFunctions.MillisecondFunctionName];
+            }
+            else if (IsDateTime(parameter.Type))
+            {
+                property = ClrCanonicalFunctions.DateTimeProperties[ClrCanonicalFunctions.MillisecondFunctionName];
+            }
+            else if (IsTimeSpan(parameter.Type))
+            {
+                property = ClrCanonicalFunctions.TimeSpanProperties[ClrCanonicalFunctions.MillisecondFunctionName];
+            }
+            else
+            {
+                property = ClrCanonicalFunctions.DateTimeOffsetProperties[ClrCanonicalFunctions.MillisecondFunctionName];
+            }
+
+            // Millisecond
+            Expression milliSecond = MakePropertyAccess(property, parameter);
+            Expression decimalMilliSecond = Expression.Convert(milliSecond, typeof(decimal));
+            Expression fractionalSeconds = Expression.Divide(decimalMilliSecond, Expression.Constant(1000m, typeof(decimal)));
+
+            return CreateFunctionCallWithNullPropagation(fractionalSeconds, arguments);
+        }
+
+        private Expression BindDateRelatedProperty(SingleValueFunctionCallNode node)
+        {
+            Expression[] arguments = BindArguments(node.Parameters);
+            Contract.Assert(arguments.Length == 1 && IsDateRelated(arguments[0].Type));
+
+            // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
+            Expression parameter = arguments[0];
+
+            PropertyInfo property;
+            if (IsDate(parameter.Type))
+            {
+                Contract.Assert(ClrCanonicalFunctions.DateProperties.ContainsKey(node.Name));
+                property = ClrCanonicalFunctions.DateProperties[node.Name];
+            }
+            else if (IsDateTime(parameter.Type))
+            {
+                Contract.Assert(ClrCanonicalFunctions.DateTimeProperties.ContainsKey(node.Name));
+                property = ClrCanonicalFunctions.DateTimeProperties[node.Name];
+            }
+            else
+            {
+                Contract.Assert(ClrCanonicalFunctions.DateTimeOffsetProperties.ContainsKey(node.Name));
+                property = ClrCanonicalFunctions.DateTimeOffsetProperties[node.Name];
+            }
+
+            return MakeFunctionCall(property, parameter);
+        }
+
+        private Expression BindTimeRelatedProperty(SingleValueFunctionCallNode node)
+        {
+            Expression[] arguments = BindArguments(node.Parameters);
+            Contract.Assert(arguments.Length == 1 && (IsTimeRelated(arguments[0].Type)));
+
+            // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
+            Expression parameter = arguments[0];
+
+            PropertyInfo property;
+            if (IsTimeOfDay(parameter.Type))
+            {
+                Contract.Assert(ClrCanonicalFunctions.TimeOfDayProperties.ContainsKey(node.Name));
+                property = ClrCanonicalFunctions.TimeOfDayProperties[node.Name];
+            }
+            else if (IsDateTime(parameter.Type))
+            {
+                Contract.Assert(ClrCanonicalFunctions.DateTimeProperties.ContainsKey(node.Name));
+                property = ClrCanonicalFunctions.DateTimeProperties[node.Name];
+            }
+            else if (IsTimeSpan(parameter.Type))
+            {
+                Contract.Assert(ClrCanonicalFunctions.TimeSpanProperties.ContainsKey(node.Name));
+                property = ClrCanonicalFunctions.TimeSpanProperties[node.Name];
+            }
+            else
+            {
+                Contract.Assert(ClrCanonicalFunctions.DateTimeOffsetProperties.ContainsKey(node.Name));
+                property = ClrCanonicalFunctions.DateTimeOffsetProperties[node.Name];
+            }
+
+            return MakeFunctionCall(property, parameter);
+        }
+
+        private Expression BindConcat(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("concat" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.Concat, arguments);
+        }
+
+        private Expression BindTrim(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("trim" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.Trim, arguments);
+        }
+
+        private Expression BindToUpper(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("toupper" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.ToUpper, arguments);
+        }
+
+        private Expression BindToLower(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("tolower" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.ToLower, arguments);
+        }
+
+        private Expression BindIndexOf(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("indexof" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.IndexOf, arguments);
+        }
+
+        private Expression BindSubstring(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("substring" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            if (arguments[0].Type != typeof(string))
+            {
+                throw new ODataException(Error.Format(SRResources.FunctionNotSupportedOnEnum, node.Name));
+            }
+
+            Expression functionCall;
+            if (arguments.Length == 2)
+            {
+                Contract.Assert(IsInteger(arguments[1].Type));
+
+                // When null propagation is allowed, we use a safe version of String.Substring(int).
+                // But for providers that would not recognize custom expressions like this, we map
+                // directly to String.Substring(int)
+                if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
+                {
+                    // Safe function is static and takes string "this" as first argument
+                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStartNoThrow, arguments);
+                }
+                else
+                {
+                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStart, arguments);
+                }
+            }
+            else
+            {
+                // arguments.Length == 3 implies String.Substring(int, int)
+                Contract.Assert(arguments.Length == 3 && IsInteger(arguments[1].Type) && IsInteger(arguments[2].Type));
+
+                // When null propagation is allowed, we use a safe version of String.Substring(int, int).
+                // But for providers that would not recognize custom expressions like this, we map
+                // directly to String.Substring(int, int)
+                if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
+                {
+                    // Safe function is static and takes string "this" as first argument
+                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStartAndLengthNoThrow, arguments);
+                }
+                else
+                {
+                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStartAndLength, arguments);
+                }
+            }
+
+            return functionCall;
+        }
+
+        private Expression BindLength(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("length" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.Length, arguments);
+        }
+
+        private Expression BindContains(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("contains" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.Contains, arguments[0], arguments[1]);
+        }
+
+        private Expression BindStartsWith(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("startswith" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.StartsWith, arguments);
+        }
+
+        private Expression BindEndsWith(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert("endswith" == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            ValidateAllStringArguments(node.Name, arguments);
+
+            Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
+
+            return MakeFunctionCall(ClrCanonicalFunctions.EndsWith, arguments);
+        }
+
+        /// <summary>
+        /// Bind function arguments
+        /// </summary>
+        /// <param name="nodes"></param>
+        /// <returns></returns>
+        protected Expression[] BindArguments(IEnumerable<QueryNode> nodes)
+        {
+            return nodes.OfType<SingleValueNode>().Select(n => Bind(n)).ToArray();
+        }
+
+        private static void ValidateAllStringArguments(string functionName, Expression[] arguments)
+        {
+            if (arguments.Any(arg => arg.Type != typeof(string)))
+            {
+                throw new ODataException(Error.Format(SRResources.FunctionNotSupportedOnEnum, functionName));
+            }
+        }
+
+
+        private Expression BindCustomMethodExpressionOrNull(SingleValueFunctionCallNode node)
+        {
+            Expression[] arguments = BindArguments(node.Parameters);
+            IEnumerable<Type> methodArgumentsType = arguments.Select(argument => argument.Type);
+
+            // Search for custom method info that are binded to the node name
+            MethodInfo methodInfo;
+            if (UriFunctionsBinder.TryGetMethodInfo(node.Name, methodArgumentsType, out methodInfo))
+            {
+                return MakeFunctionCall(methodInfo, arguments);
+            }
+
+            return null;
+        }
+        private Type RetrieveClrTypeForConstant(IEdmTypeReference edmTypeReference, ref object value)
+        {
+            Type constantType = EdmLibHelpers.GetClrType(edmTypeReference, Model, InternalAssembliesResolver);
+
+            if (value != null && edmTypeReference != null && edmTypeReference.IsEnum())
+            {
+                ODataEnumValue odataEnumValue = (ODataEnumValue)value;
+                string strValue = odataEnumValue.Value;
+                Contract.Assert(strValue != null);
+
+                constantType = Nullable.GetUnderlyingType(constantType) ?? constantType;
+                value = Enum.Parse(constantType, strValue);
+            }
+
+            if (edmTypeReference != null &&
+                edmTypeReference.IsNullable &&
+                (edmTypeReference.IsDate() || edmTypeReference.IsTimeOfDay()))
+            {
+                constantType = Nullable.GetUnderlyingType(constantType) ?? constantType;
+            }
+
+            return constantType;
         }
     }
 }
