+++ test/UnitTest/Microsoft.AspNet.OData.Test.Shared/Query/ApplyQueryOptionTest.cs
@@ -46,6 +46,13 @@ namespace Microsoft.AspNet.OData.Test.Query
                         }
                     },
                     {
+                        "aggregate(cast(CustomerId, Edm.Int64) with sum as CustomerId)",
+                        new List<Dictionary<string, object>>
+                        {
+                            new Dictionary<string, object> { { "CustomerId", 15L} }
+                        }
+                    },
+                    {
                         "aggregate(SharePrice with sum as SharePrice)",
                         new List<Dictionary<string, object>>
                         {
